+++ universal-login-commons/lib/index.ts
@@ -1,6 +1,6 @@
 export {sleep, waitToBeMined, waitUntil, waitForContractDeploy} from './utils/wait';
 export {debounce} from './utils/debounce';
-export {Procedure, Predicate, Partial} from './utils/types';
+export {Procedure, Predicate, Partial, Message} from './utils/types';
 export {copy} from './utils/copy';
 export {SuggestionsService} from './services/SuggestionsService';
 export {WalletExistenceVerifier, WalletSelectionService} from './services/WalletSelectionService';
